Started on Holy Strike customization
Will be adding a custom texture and rarity, as well as a custom block with a custom UI as a way to obtain the enchantment.

I will also look more into how an enchantment's context works to facilitate the mod creation as my current way to trigger it is very complicated and I may not be able to finish it.